pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "inquisitor-framework"
version = "0.1.0"
description = "A manifest-driven, protocol-based framework for automated inquiry and analysis."
readme = "MANIFEST.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Sheewi", email = "sheewi.tech@gmail.com" }
]
keywords = ["analysis", "automation", "framework", "manifest", "research"]

dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "aiofiles>=23.0.0",
    "tenacity>=8.2.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "networkx>=3.0",
    "torch>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "redis>=5.0.0",
    "prometheus-client>=0.17.0",
    "opentelemetry-api>=1.19.0",
    "opentelemetry-sdk>=1.19.0",
]

[project.scripts]
inquisitor = "inquisitor.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/inquisitor"]

[tool.pytest.ini_options]
addopts = "--cov=src/inquisitor --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["src"]
